Saving now works right.  Enabling now works right.  Had to make several GUI changes to make it functional.
Save button will now set itself to the first responder before saving, which should cause any edited text to get dumped to the Array controller. The typical behavior being that until you leave the text box, the text hasn't been committed yet.  Moving the responder seems to get the job done.

Enable/Disable is now handled entirely by the saveConfig method.  That means I had to move the host controller calls back to the main controller, but that's not a big deal.

Passenger Controller is now the delegate for the textfields; the enabled checkbox and the prod/dev matrix, have a separate method that gets called when they're activated that just fires the delegate.  That lets me set a flag for edits, so I can display a reminder to save.